backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
google-cloud-aiplatform
google-cloud-storage
google-cloud-vision
google-cloud-discoveryengine
google-adk
nest-asyncio
python-multipart
aiofiles
python-dotenv
httpx
pypdf
pdfplumber
python-docx
Pillow
python-dateutil
pytest
pytest-cov
